Removed DAO interfaces as their job will be handled to ApiProvider. Moved classes from dao.object to dao package.
