--- app/src/main/java/com/astronauts/astronaut/dao/ConversationListDAO.java
@@ -1,9 +1,0 @@
-package com.astronauts.astronaut.dao;
-
-/**
- * Created by end on 05.05.17.
- */
-
-public interface ConversationListDAO {
-
-}
