--- app/src/main/java/com/astronauts/astronaut/dao/DigDAO.java
@@ -1,8 +1,0 @@
-package com.astronauts.astronaut.dao;
-
-/**
- * Created by end on 04.05.17.
- */
-
-public interface DigDAO {
-}
