+++ app/src/main/java/com/astronauts/astronaut/dao/EntryComment.java
@@ -1,4 +1,4 @@
-package com.astronauts.astronaut.dao.object;
+package com.astronauts.astronaut.dao;
 
 import java.net.URI;
 import java.util.Date;
