+++ app/src/main/java/com/astronauts/astronaut/dao/Profile.java
@@ -1,4 +1,4 @@
-package com.astronauts.astronaut.dao.object;
+package com.astronauts.astronaut.dao;
 
 import java.net.URI;
